add osmo_gsup_set_reply()
As I am implementing GSUP proxy forwarding between osmo-hlr instances, it
becomes apparent that many code paths omit to copy back essential information
in replies to GSUP requests:

If the response to a request omits to include an identical cn_domain indicator,
proxy routing becomes hard.

Other fields, like IMSI and session_id, also need to be identical in the
response.

Add this API that makes sure these commonly known fields are copied from a GSUP
request to the response, so that we can centrally add future fields here. After
this, it is up to osmo-hlr to use it in all relevant GSUP message composition.

Change-Id: Id9692880079ea0f219f52d81b1923a76fc640566