initial support for static userspace probes via systemtap
This adds a --enable-systemtap configure option, which will then
add static tracepoints to the generated libosmocore binary.

At this point, only two tracepoints are supported: stderr_start
and stderr_done.  They can be used to trace the amount of time
a libosmocore-using application spends in blocking calls to log
to stderr.

Change-Id: I7e1ab664241deb524c9582cbd1bec31af46c747e